inverter aus UI installieren

+++ functions/storm.house.bash
@@ -303,7 +303,9 @@ setup_wb_config() {
     if [[ -z "${1:-$wallboxtype}" ]]; then echo "SKIPPED (no wallbox defined)"; return 1; fi
   fi
 
-   if ! cond_redirect install -m 755 "${sdIncludesDir}/setup_inverter" /usr/local/sbin; then echo "FAILED (install setup_inverter)"; return 1; fi
+  if [[ ! -f /usr/local/sbin/setup_inverter ]]; then
+    if ! cond_redirect install -m 755 "${sdIncludesDir}/setup_inverter" /usr/local/sbin; then echo "FAILED (install setup_inverter)"; return 1; fi
+  fi
 
   for component in things items rules; do
     rm -f "${OPENHAB_CONF:-/etc/openhab}/${component}/wb.${component}"
